change path for Header component and rename Default layout to UserHomePageLayout
