+++ client/src/layouts/ManageMyNetworkLayout/ManageMyNetworkLayout.js
@@ -3,7 +3,7 @@ import { Box, Typography } from '@mui/material';
 import { Container } from '@mui/material';
 import Grid from '@mui/material/Grid';
 import Advertising from '../../components/Advertising/Advertising';
-import Header from '../DefaultLayouts/Header/Header';
+import Header from '../UserHomePageLayout/Header/Header';
 import { mobileScreen } from '../../components/Theme/Theme';
 
 function ManageMyNetworkLayout({ children }) {
@@ -19,12 +19,12 @@ function ManageMyNetworkLayout({ children }) {
                 }}
             >
                 <Grid container rowSpacing={3} columnSpacing={{ xs: 1, sm: 2, md: 3 }}>
-                    <Grid item xs={12} sm={7} md={8}>
+                    <Grid item xs={12} sm={7} md={8} lg={9}>
                         <Box sx={{ minHeight: '10vh', borderRadius: '24px' }}>
                             <Box>{children}</Box>
                         </Box>
                     </Grid>
-                    <Grid item xs={12} sm={5} md={4}>
+                    <Grid item xs={12} sm={5} md={4} lg={3}>
                         {/* Manage my network */}
                         <Box sx={{ position: 'relative' }}>
                             <Advertising />
